+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Latest
+
+New APIs
+- earth2grid.healpix
+  - zonal_average
+  - ring2double
+  - to_rotated_pixelization
+  - to_double_pixelization
+
 ## 2025.4.1
 
 Breaking changes: